+++ src/containers/Feed/style.js
@@ -10,37 +10,6 @@ export const PostContainer = styled.div`
   background-color: #3D4F52;
   font-family: "Ubuntu";
 `
-export const NewPostContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin: 20px;
-  background-color: #fff;
-  width: 500px;
-  padding: 10px;
-  box-shadow: 10px 10px 10px #2A3439;
-`
-export const StyledTextArea = styled(TextField)`
-  width: 100%;
-`
-export const StyledTitle = styled.h2`
-  margin: 0;
-`
-export const ButtonPost = styled.button`
-  width: 60%;
-  border: none;
-  font-family: 'Ubuntu';
-  height: 40px;
-  cursor: pointer;
-  font-weight: bold;
-  transition: all 0.5s;
-
-  &:hover{
-    background-color: #E67221;
-    color: #fff;
-  }
-`
 export const PostCard = styled.div`
   width: 450px;
   height: 250px;
